Added documentation for service
