+++ Sources/ForageSDK/Foundation/Network/LiveForageService.swift
@@ -8,48 +8,6 @@
 import Foundation
 import VGSCollectSDK
 
-internal protocol ForageService: AnyObject {
-    var provider: Provider { get }
-    
-    func tokenizeEBTCard(
-        request: ForagePANRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void) -> Void
-    
-    func getXKey(
-        bearerToken: String,
-        completion: @escaping (Result<ForageXKeyModel, Error>) -> Void) -> Void
-    
-    func getBalance(
-        request: ForageRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void) -> Void
-    
-    func retrieveCheckBalance(
-        request: ForageRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func requestCapturePayment(
-        request: ForageRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func retrieveCapturedPayment(
-        request: ForageRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func cancelRequest()
-}
-
-internal protocol Polling: AnyObject {
-    func polling(
-        response: VGSResponse,
-        request: ForageRequestModel,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func pollingMessage(
-        message: MessageResponseModel,
-        request: ForageRequestModel,
-        completion: @escaping (Result<MessageResponseModel, Error>) -> Void) -> Void
-}
-
 internal class LiveForageService: ForageService {
     
     // MARK: Properties
@@ -58,6 +16,7 @@ internal class LiveForageService: ForageService {
     private var collector: VGSCollect?
     private var maxAttempts: Int = 10
     private var intervalBetweenAttempts: Double = 1.0
+    private var retryCount = 0
     
     init(_ collector: VGSCollect?) {
         self.collector = collector
@@ -65,6 +24,11 @@ internal class LiveForageService: ForageService {
     
     // MARK: Tokenize EBT card
     
+    /// Tokenize a given *ForagePANRequestModel* object
+    ///
+    /// - Parameters:
+    ///  - request: *ForagePANRequestModel* contains ebt card object.
+    ///  - completion: Returns tokenized object.
     internal func tokenizeEBTCard(request: ForagePANRequestModel, completion: @escaping (Result<Data?, Error>) -> Void) {
         do { try provider.execute(endpoint: ForageAPI.panNumber(request: request), completion: completion) }
         catch { completion(.failure(error)) }
@@ -72,6 +36,11 @@ internal class LiveForageService: ForageService {
     
     // MARK: X-key
     
+    /// Retrieve from ForageAPI the X-Key header to perform request.
+    ///
+    /// - Parameters:
+    ///  - bearerToken: Session authorization token.
+    ///  - completion: Returns *ForageXKeyModel* object.
     internal func getXKey(bearerToken: String, completion: @escaping (Result<ForageXKeyModel, Error>) -> Void) {
         do { try provider.execute(model: ForageXKeyModel.self, endpoint: ForageAPI.xKey(bearerToken: bearerToken), completion: completion) }
         catch { completion(.failure(error)) }
@@ -79,10 +48,16 @@ internal class LiveForageService: ForageService {
     
     // MARK: Cancel request
     
+    /// Cancel any ongoing request.
     internal func cancelRequest() { provider.stopRequestOnGoing() }
     
     // MARK: Check balance
     
+    /// Perform VGS SDK request to retrieve balance.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns balance object.
     internal func getBalance(
         request: ForageRequestModel,
         completion: @escaping (Result<Data?, Error>) -> Void) -> Void
@@ -114,6 +89,11 @@ internal class LiveForageService: ForageService {
             }
     }
     
+    /// Perform request to Forage API after polling got message success.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns balance object.
     internal func retrieveCheckBalance(request: ForageRequestModel,
                                        completion: @escaping (Result<Data?, Error>) -> Void) {
         do {
@@ -125,6 +105,11 @@ internal class LiveForageService: ForageService {
     
     // MARK: Capture payment
     
+    /// Perform VGS SDK request to capture payment.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns captured payment object.
     internal func requestCapturePayment(
         request: ForageRequestModel,
         completion: @escaping (Result<Data?, Error>) -> Void)
@@ -156,6 +141,11 @@ internal class LiveForageService: ForageService {
             }
     }
     
+    /// Perform request to Forage API after polling got message success.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns captured payment object.
     internal func retrieveCapturedPayment(
         request: ForageRequestModel,
         completion: @escaping (Result<Data?, Error>) -> Void)
@@ -168,7 +158,15 @@ internal class LiveForageService: ForageService {
 // MARK: - Polling
 
 extension LiveForageService: Polling {
+    /// Process VGSData for polling message.
+    ///
+    /// - Parameters:
+    ///  - response: The *VGSResponse* which contains the result from VGS SDK.
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Which will return the result.
     internal func polling(response: VGSResponse, request: ForageRequestModel, completion: @escaping (Result<Data?, Error>) -> Void) {
+        retryCount = 0
+        
         switch response {
         case .success(_, let data, _):
             provider.processVGSData(model: MessageResponseModel.self, code: nil, data: data, response: nil) { [weak self] messageResponse in
@@ -197,48 +195,59 @@ extension LiveForageService: Polling {
         }
     }
     
+    /// Polls message to check payment status
+    ///
+    /// - Parameters:
+    ///  - message: The *MessageResponseModel* which contains the message.
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Which will return the message for another retry or success.
     internal func pollingMessage(
         message: MessageResponseModel,
         request: ForageRequestModel,
         completion: @escaping (Result<MessageResponseModel, Error>) -> Void) -> Void
     {
-        var retryCount = 0
         do {
             try provider.execute(model: MessageResponseModel.self, endpoint: ForageAPI.message(request: message, bearerToken: request.authorization, merchantID: request.merchantID), completion: { [weak self] result in
                 guard let self = self else { return }
                 switch result {
                 case .success(let data):
+                    /// check message is not failed and is completed
                     if data.failed == false && data.status == .completed {
-                        autoreleasepool {
-                            completion(.success(data))
-                        }
+                        completion(.success(data))
+                    /// check message is failed to return error immediately
                     } else if data.failed == true {
-                        autoreleasepool {
-                            completion(.failure(NSError(domain: "Message failed", code: 001, userInfo: nil)))
-                        }
-                    } else if retryCount < self.maxAttempts {
-                        retryCount += 1
-                        DispatchQueue.main.asyncAfter(deadline: .now() + self.intervalBetweenAttempts) {
+                        completion(.failure(NSError(domain: "Message failed", code: 001, userInfo: nil)))
+                    /// check maxAttempts to retry
+                    } else if self.retryCount < self.maxAttempts {
+                        self.waitNextAttempt {
                             self.pollingMessage(
                                 message: data,
                                 request: request,
                                 completion: completion
                             )
                         }
+                    /// in case run out of attempts
                     } else {
-                        autoreleasepool {
-                            completion(.failure(NSError(domain: "Reached max attempts", code: 002, userInfo: nil)))
-                        }
+                        completion(.failure(NSError(domain: "Reached max attempts", code: 429, userInfo: nil)))
                     }
                     
                 case .failure(let error):
-                    autoreleasepool {
-                        completion(.failure(error))
-                    }
+                    completion(.failure(error))
                 }
             })
         } catch {
             completion(.failure(error))
+        }
+    }
+    
+    /// Support function to update retry count and interval between attempts.
+    ///
+    /// - Parameters:
+    ///  - completion: Which will return after an wait.
+    private func waitNextAttempt(completion: @escaping () -> ()) {
+        retryCount = retryCount + 1
+        DispatchQueue.main.asyncAfter(deadline: .now() + self.intervalBetweenAttempts) {
+            completion()
         }
     }
 }
